refactor: Move main function to the class "PasswordCommand"
